chore(query): make char function compatible with pg/sn
