+++ src/query/functions/src/scalars/binary.rs
@@ -13,37 +13,25 @@
 // limitations under the License.
 
 use std::io::Write;
-use std::sync::Arc;
 
 use databend_common_expression::error_to_null;
-use databend_common_expression::passthrough_nullable;
 use databend_common_expression::types::binary::BinaryColumn;
 use databend_common_expression::types::binary::BinaryColumnBuilder;
 use databend_common_expression::types::nullable::NullableColumn;
 use databend_common_expression::types::string::StringColumn;
 use databend_common_expression::types::string::StringColumnBuilder;
-use databend_common_expression::types::AnyType;
 use databend_common_expression::types::BinaryType;
 use databend_common_expression::types::Bitmap;
 use databend_common_expression::types::BitmapType;
-use databend_common_expression::types::DataType;
 use databend_common_expression::types::GeographyType;
 use databend_common_expression::types::GeometryType;
-use databend_common_expression::types::NumberDataType;
 use databend_common_expression::types::NumberType;
 use databend_common_expression::types::StringType;
-use databend_common_expression::types::UInt8Type;
 use databend_common_expression::types::VariantType;
 use databend_common_expression::vectorize_1_arg;
-use databend_common_expression::Column;
 use databend_common_expression::EvalContext;
-use databend_common_expression::Function;
 use databend_common_expression::FunctionDomain;
-use databend_common_expression::FunctionEval;
-use databend_common_expression::FunctionFactory;
 use databend_common_expression::FunctionRegistry;
-use databend_common_expression::FunctionSignature;
-use databend_common_expression::Scalar;
 use databend_common_expression::Value;
 
 pub fn register(registry: &mut FunctionRegistry) {
@@ -215,54 +203,6 @@ pub fn register(registry: &mut FunctionRegistry) {
         |_, _| FunctionDomain::Full,
         error_to_null(eval_from_base64),
     );
-
-    let char = FunctionFactory::Closure(Box::new(|_, args_type: &[DataType]| {
-        if args_type.is_empty() {
-            return None;
-        }
-        let has_null = args_type.iter().any(|t| t.is_nullable_or_null());
-        let f = Function {
-            signature: FunctionSignature {
-                name: "char".to_string(),
-                args_type: vec![DataType::Number(NumberDataType::UInt8); args_type.len()],
-                return_type: DataType::Binary,
-            },
-            eval: FunctionEval::Scalar {
-                calc_domain: Box::new(|_, _| FunctionDomain::Full),
-                eval: Box::new(char_fn),
-            },
-        };
-
-        if has_null {
-            Some(Arc::new(f.passthrough_nullable()))
-        } else {
-            Some(Arc::new(f))
-        }
-    }));
-    registry.register_function_factory("char", char);
-
-    let nullable_char = FunctionFactory::Closure(Box::new(|_, args_type: &[DataType]| {
-        if args_type.is_empty() {
-            return None;
-        }
-        Some(Arc::new(Function {
-            signature: FunctionSignature {
-                name: "char".to_string(),
-                args_type: vec![
-                    DataType::Nullable(Box::new(DataType::Number(
-                        NumberDataType::UInt8
-                    )));
-                    args_type.len()
-                ],
-                return_type: DataType::Nullable(Box::new(DataType::Binary)),
-            },
-            eval: FunctionEval::Scalar {
-                calc_domain: Box::new(|_, _| FunctionDomain::MayThrow),
-                eval: Box::new(passthrough_nullable(char_fn)),
-            },
-        }))
-    }));
-    registry.register_function_factory("char", nullable_char);
 }
 
 fn eval_binary_to_string(val: Value<BinaryType>, ctx: &mut EvalContext) -> Value<StringType> {
@@ -352,33 +292,5 @@ pub fn vectorize_string_to_binary(
 
             Value::Column(builder.build())
         }
-    }
-}
-
-fn char_fn(args: &[Value<AnyType>], _: &mut EvalContext) -> Value<AnyType> {
-    let args = args
-        .iter()
-        .map(|arg| arg.try_downcast::<UInt8Type>().unwrap())
-        .collect::<Vec<_>>();
-
-    let len = args.iter().find_map(|arg| match arg {
-        Value::Column(col) => Some(col.len()),
-        _ => None,
-    });
-    let input_rows = len.unwrap_or(1);
-
-    let mut builder = BinaryColumnBuilder::with_capacity(input_rows, 0);
-
-    for _ in 0..input_rows {
-        for arg in &args {
-            let val = arg.index(0).unwrap();
-            builder.put_u8(val);
-        }
-        builder.commit_row();
-    }
-
-    match len {
-        Some(_) => Value::Column(Column::Binary(builder.build())),
-        _ => Value::Scalar(Scalar::Binary(builder.build_scalar())),
     }
 }
